ParsingSite/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
Pillow==7.1.2
pytesseract==0.3.4
requests==2.23.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.9
